split functions into separate scripts
